AgenticAI/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
ollama
huggingface-hub
sentence_transformers
google-generativeai
numpy==1.26.4